Finished working on the "header" block and the burger-menu button.
html - The "header" block has been redesigned. Additional content, new classes and tags were added.

css - Added scss/css style properties for header block, its classes and tags. Animation for burger menu (keyframes) was fully animated. Flex-elements in the code were used.

assets\header, the content for the burger-menu was added to the folder assets\header.

js - the function "hb.addEventListener" was added to create the animation of the burger-menu. When clicked, it activates the animation and blocks the surrounding background, allowing you to scroll through the menu separately.

@AidenOrtiz