Development work was started on the content-main block. In the course of the work, the block was slightly redesigned, and all the work on it is now complete.
html - the content-main block has been added. Content, tags and classes were added to it for further work. The block was slightly reworked in the course of the work.

css - the "content-main" block and its classes were added to the style.scss/css file. Styles properties for the block, its tags and classes were applied. A small animation was added to class "conten-main__menu" and flex-elements were used.

assets - added material for further work with block into assets\content-main folder.

@AidenOrtiz