The"content-bottom" block and "scroll.js" scripts have been started and finished.
html - The "content-bottom" block has been added. Tags, content and classes were added to it, to further work with the content.

css - The "content-bottom" block and its classes were added to the style.scss/css file. Appropriate style properties were applied, and flex elements and animations were used.

assets - material for working with the block has been added to the assets\content-bottom folder.

js - in the file scroll.js, was created a script which is responsible for smooth transition from the bottom page of the site (the button is an arrow in a circle) to the site header.