Work has started on the "footer" block.
html - added a "footer" block and a class of the same name. Additional classes for the block, tags and added content have been added.

css - the footer block and its classes were added to the style.scss/css file.

assets - material for working with block was added to assets\footer folder.

@AidenOrtiz