Work on the project "origami-caffe", completed.
html - In the class "adress-image", replaced the picture, the API map of Yandex, for more convenient and beautiful use. Slightly redone some of the classes, added additional tags.

css - The site font was re-imported, the file was used to zero styles. Additional "hover" was added for some elements on the site, comments were added to the scss/css file, for easier orientation when working. In the class "drinksprice4", additional indentation for prices has been added.

@AidenOrtiz